refactor: 지역 애플리케이션 레이어 구조 변경 (#210)
