+++ bootstrap/api/src/main/kotlin/kr/wooco/woocobe/api/region/RegionController.kt
@@ -3,13 +3,13 @@ package kr.wooco.woocobe.api.region
 import kr.wooco.woocobe.api.region.request.CreatePreferenceRegionRequest
 import kr.wooco.woocobe.api.region.response.CreatePreferenceRegionResponse
 import kr.wooco.woocobe.api.region.response.PreferenceDetailResponse
-import kr.wooco.woocobe.core.region.application.command.port.`in`.AddPreferenceRegionUseCase
-import kr.wooco.woocobe.core.region.application.command.port.`in`.DeletePreferenceRegionCommand
-import kr.wooco.woocobe.core.region.application.command.port.`in`.DeletePreferenceRegionUseCase
-import kr.wooco.woocobe.core.region.application.query.port.`in`.GetAllPreferenceRegionQuery
-import kr.wooco.woocobe.core.region.application.query.port.`in`.GetAllPreferenceRegionUseCase
-import kr.wooco.woocobe.core.region.application.query.port.`in`.GetPreferenceRegionQuery
-import kr.wooco.woocobe.core.region.application.query.port.`in`.GetPreferenceRegionUseCase
+import kr.wooco.woocobe.core.region.application.port.`in`.AddPreferenceRegionUseCase
+import kr.wooco.woocobe.core.region.application.port.`in`.DeletePreferenceRegionCommand
+import kr.wooco.woocobe.core.region.application.port.`in`.DeletePreferenceRegionUseCase
+import kr.wooco.woocobe.core.region.application.port.`in`.GetAllPreferenceRegionQuery
+import kr.wooco.woocobe.core.region.application.port.`in`.GetAllPreferenceRegionUseCase
+import kr.wooco.woocobe.core.region.application.port.`in`.GetPreferenceRegionQuery
+import kr.wooco.woocobe.core.region.application.port.`in`.GetPreferenceRegionUseCase
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.DeleteMapping
