+++ core/src/main/kotlin/kr/wooco/woocobe/core/region/application/port/in/AddPreferenceRegionUseCase.kt
@@ -1,4 +1,4 @@
-package kr.wooco.woocobe.core.region.application.command.port.`in`
+package kr.wooco.woocobe.core.region.application.port.`in`
 
 import kr.wooco.woocobe.core.region.domain.entity.PreferenceRegion
